# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestReviews API - Restoranlar va Sharhlar",
        "description": "Mobile ilova uchun REST API. Restoranlarni kashf qilish, sharh qoldirish, sevimlilar va boshqalar.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gastronomic.webclub.uz",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://restreviews.test",
            "description": "Laragon Server"
        }
    ],
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "❤️ Sevimlilar"
                ],
                "summary": "Sevimlilar ro'yxati",
                "description": "Login qilgan foydalanuvchining barcha sevimli restoranlar ro'yxati",
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en). Default: kk",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "kk",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorites list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/restaurants/{id}/favorite": {
            "post": {
                "tags": [
                    "❤️ Sevimlilar"
                ],
                "summary": "Sevimlilar ro'yxatiga qo'shish/o'chirish",
                "description": "Restoranni sevimlilar ro'yxatiga qo'shish yoki o'chirish. Faqat autentifikatsiya qilingan foydalanuvchilar uchun.",
                "operationId": "7007d8d312a4e78f3e05186adb89a117",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Restoran ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sevimlilar holati o'zgartirildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sevimlilar ro'yxatiga qo'shildi"
                                        },
                                        "data": {
                                            "properties": {
                                                "is_favorited": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya xatosi"
                    },
                    "404": {
                        "description": "Restoran topilmadi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/favorites/map": {
            "get": {
                "tags": [
                    "❤️ Sevimlilar"
                ],
                "summary": "Sevimli restoranlarni xaritada ko'rsatish",
                "description": "Client sevimli restoranlarining koordinatalari xarita uchun",
                "operationId": "17012642d364a1329a23a2d6c7a8d434",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en). Default: kk",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "kk",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sevimli restoranlar koordinatalari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Autentifikatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}